requirements.txt
openai-agents>=0.1.0
python-telegram-bot[job-queue]>=21.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
sqlalchemy[asyncio]>=2.0
alembic
asyncpg
psycopg2-binary
python-dotenv
pydantic>=2.0
pytz
mercadopago>=2.0.0
aiohttp>=3.9.0
